Change PlaceholderStyle from private to public
