+++ Kwartracker/Scenes/Sign-in Page/Content/Sign-in Reusables/UserField.swift
@@ -70,20 +70,3 @@ private extension UserField {
         }
     }
 }
-
-private struct PlaceholderStyle: ViewModifier {
-    var showPlaceHolder: Bool
-    var placeholder: String
-
-    public func body(content: Content) -> some View {
-        ZStack(alignment: .leading) {
-            if showPlaceHolder {
-                Text(placeholder)
-                    .italic()
-                    .foregroundColor(.secondary)
-            }
-            content
-                .foregroundColor(Color.black)
-        }
-    }
-}
